feat: fock space flow
Implement flow as working in Fock space, that is, to individually transform coordinates of different
states with different energy quantum numbers. The jacobians are defined w.r.t within the same
quantum number, for example, for state i, the jacobian factor that is multiplied to wave function
only takes partial fi/ partial xi where fi is the i-th coordinate that transformed from flow and xi
is the original coordinate of i-th state.

BREAKING CHANGE: Flow working in Fock Space